Added option to select vertex/edge thickness parameters in PeriodicExploration. Fixed unsigned/signed comparison compiler warning.

+++ tools/Wires/Parameters/ParameterList.h
@@ -30,7 +30,7 @@ class ParameterList {
         }
 
         void set_dofs(const VectorF& dofs) {
-            assert(dofs.size() == m_params.size());
+            assert(size_t(dofs.size()) == m_params.size());
             size_t count = 0;
             for (auto param : m_params) {
                 param->set_value(dofs[count]);
